Split Linter.hs out to Tools.hs and Checks.hs
