+++ src/Main.hs
@@ -1,10 +1,15 @@
+{-# LANGUAGE RecordWildCards #-}
+
+
 module Main where
 
 import           Data.Foldable
 import           System.Environment
 
+import           Nix.Expr.Types.Annotated (stripAnnotation)
 import           Nix.Parser
 import           Nix.Pretty
+import           Nix.Render
 
 import           Nix.Linter
 import           Nix.Linter.Types
@@ -12,8 +17,19 @@ import           Nix.Linter.Types
 main :: IO ()
 main = do
   args <- getArgs
-  for_ args $ \arg -> do
-    parsed <- parseNixFileLoc arg
+  for_ args (runCheck checkAll)
+
+
+runCheck :: Check -> FilePath -> IO ()
+runCheck check file = do
+    parsed <- parseNixFileLoc file
     case parsed of
-      Success parse -> for_ (checkAll parse) (putStrLn . prettyOffense)
-      Failure err   -> putStr "Parse Error: " >> print  err
+      Success parse -> for_ (check parse) $ \x@(Offense {..}) -> do
+        let explanation = prettyOffense x
+        putStrLn explanation
+        case rewrite of
+          Nothing -> pure ()
+          (Just e) -> do
+            putStrLn "Why not:"
+            print $ prettyNix e
+      Failure err   -> putStr "Parse Error: " >> print err
